[Sprites] Extracts bonding box generation into RenderSprite. Adding some simple caching to prevent needless calculations.
