+++ sources/engine/Xenko.Engine/Rendering/Sprites/SpriteRenderProcessor.cs
@@ -33,33 +33,7 @@ namespace Xenko.Rendering.Sprites
 
                 if (renderSprite.Enabled)
                 {
-                    var transform = renderSprite.TransformComponent;
-
-                    // TODO GRAPHICS REFACTOR: Proper bounding box. Reuse calculations in sprite batch.
-                    // For now we only set a center for sorting, but no extent (which disable culling)
-                    renderSprite.BoundingBox = new BoundingBoxExt { Center = transform.WorldMatrix.TranslationVector };
-                    renderSprite.RenderGroup = renderSprite.SpriteComponent.RenderGroup;
-
-                    // update the sprite bounding box
-                    Vector3 halfBoxSize;
-                    var halfSpriteSize = currentSprite?.Size / 2 ?? Vector2.Zero;
-                    var worldMatrix = renderSprite.TransformComponent.WorldMatrix;
-                    var boxOffset = worldMatrix.TranslationVector;
-                    if (renderSprite.SpriteComponent.SpriteType == SpriteType.Billboard)
-                    {
-                        // Make a gross estimation here as we don't have access to the camera view matrix
-                        // TODO: move this code or grant camera view matrix access to this processor
-                        var maxScale = Math.Max(worldMatrix.Row1.Length(), Math.Max(worldMatrix.Row2.Length(), worldMatrix.Row3.Length()));
-                        halfBoxSize = maxScale * halfSpriteSize.Length() * Vector3.One;
-                    }
-                    else
-                    {
-                        halfBoxSize = new Vector3(
-                            Math.Abs(worldMatrix.M11 * halfSpriteSize.X + worldMatrix.M21 * halfSpriteSize.Y),
-                            Math.Abs(worldMatrix.M12 * halfSpriteSize.X + worldMatrix.M22 * halfSpriteSize.Y),
-                            Math.Abs(worldMatrix.M13 * halfSpriteSize.X + worldMatrix.M23 * halfSpriteSize.Y));
-                    }
-                    renderSprite.BoundingBox = new BoundingBoxExt(boxOffset - halfBoxSize, boxOffset + halfBoxSize);
+                    renderSprite.CalculateBoundingBox();
                 }
 
                 // TODO Should we allow adding RenderSprite without a CurrentSprite instead? (if yes, need some improvement in RenderSystem)
